tools.manipulator.SetOfAtomContainersManipulator.getTotalFormalCharge() returns double instead of int. The multiplier is now considered.
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@2807 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/tools/manipulator/SetOfMoleculesManipulator.java
@@ -75,7 +75,7 @@ public class SetOfMoleculesManipulator {
 	/**
 	 * @see SetOfAtomContainersManipulator
 	 */
-	public static int getTotalFormalCharge(SetOfMolecules set) {
+	public static double getTotalFormalCharge(SetOfMolecules set) {
 		return SetOfAtomContainersManipulator.getTotalFormalCharge(set);
 	}
